move `reduction` into `parse`
